refactor: 拆分警报记录与通知模块
