+++ app.py
@@ -38,6 +38,7 @@ from goldmonitor import review_notes as review_notes_core
 from goldmonitor import risk_analysis as risk_analysis_core
 from goldmonitor import scheduler as scheduler_core
 from goldmonitor import settings_store as settings_store_core
+from goldmonitor import socket_alert_log as socket_alert_log_core
 from goldmonitor import socket_alert_rules as socket_alert_rules_core
 from goldmonitor import socket_history_review as socket_history_review_core
 from goldmonitor import socket_operations as socket_operations_core
@@ -6836,74 +6837,28 @@ socket_history_review_core.register_history_review_handlers(
 )
 
 
-@socketio.on("export_alert_log")
-def on_export_alert_log():
-    filename = f"GoldMonitor-alert-log-{datetime.now().strftime('%Y%m%d-%H%M%S')}.csv"
-    try:
-        content, count = build_alert_log_csv()
-        saved_path = save_export_file(filename, content)
-        emit("alert_log_exported", {
-            "ok": True,
-            "filename": filename,
-            "saved_path": saved_path,
-            "count": count,
-        })
-    except OSError as exc:
-        emit("alert_log_export_error", build_export_error_payload(f"告警记录导出失败: {exc}"))
-
-
-@socketio.on("clear_alert_log")
-def on_clear_alert_log():
-    ok = clear_alert_log_archive()
-    if ok:
-        alert_log.clear()
-    socketio.emit("alert_log_cleared", {"ok": ok})
-
-
-@socketio.on("update_alert_log_status")
-def on_update_alert_log_status(data=None):
-    if not isinstance(data, dict):
-        emit("alert_log_status_error", {"message": "告警记录状态参数无效"})
-        return
-    ok, entry = update_alert_log_status(
-        data.get("id"),
-        read=data.get("read") if "read" in data else None,
-        acknowledged=data.get("acknowledged") if "acknowledged" in data else None,
-    )
-    if not ok:
-        emit("alert_log_status_error", {"message": "未找到对应告警记录"})
-        return
-    socketio.emit("alert_log_status_updated", {"ok": True, "entry": entry})
-
-
-@socketio.on("update_alert_log_handling")
-def on_update_alert_log_handling(data=None):
-    if not isinstance(data, dict):
-        emit("alert_log_handling_error", {"message": "告警处理参数无效"})
-        return
-    ok, entry = update_alert_log_handling(
-        data.get("id"),
-        handled=data.get("handled") if "handled" in data else None,
-        note=data.get("note") if "note" in data else None,
-    )
-    if not ok:
-        emit("alert_log_handling_error", {"message": "未找到对应告警记录"})
-        return
-    socketio.emit("alert_log_handling_updated", {"ok": True, "entry": entry})
-
-
-@socketio.on("resend_alert_notification")
-def on_resend_alert_notification(data=None):
-    if not isinstance(data, dict):
-        emit("alert_notification_resend_error", {"message": "告警通知重发参数无效"})
-        return
-    ok, entry = resend_alert_notification(data.get("id"), start_delivery=False)
-    if not ok:
-        emit("alert_notification_resend_error", {"message": "未找到对应告警记录"})
-        return
-    socketio.emit("alert_notification_resent", {"ok": True, "entry": entry})
-    _start_alert_notification_delivery(entry, _alert_resend_title(entry))
-
+socket_alert_log_core.register_alert_log_handlers(
+    socketio,
+    now_factory=lambda: datetime.now(),
+    build_alert_log_csv=lambda: build_alert_log_csv(),
+    save_export_file=lambda filename, content: save_export_file(filename, content),
+    build_export_error_payload=lambda message: build_export_error_payload(message),
+    clear_alert_log_archive=lambda: clear_alert_log_archive(),
+    clear_alert_log_memory=lambda: alert_log.clear(),
+    update_alert_log_status=(
+        lambda alert_id, **kwargs: update_alert_log_status(alert_id, **kwargs)
+    ),
+    update_alert_log_handling=(
+        lambda alert_id, **kwargs: update_alert_log_handling(alert_id, **kwargs)
+    ),
+    resend_alert_notification=(
+        lambda alert_id, **kwargs: resend_alert_notification(alert_id, **kwargs)
+    ),
+    start_alert_notification_delivery=(
+        lambda entry, title: _start_alert_notification_delivery(entry, title)
+    ),
+    alert_resend_title=lambda entry: _alert_resend_title(entry),
+)
 
 
 # ---------- 共享状态 (托盘与窗口通信) ----------
