+++ templates/index.html
@@ -679,6 +679,7 @@
 <script src="/static/risk-analysis-center.js?v={{ app_version }}"></script>
 <script src="/static/alert-rule-center.js?v={{ app_version }}"></script>
 <script src="/static/portfolio-center.js?v={{ app_version }}"></script>
+<script src="/static/alert-log-center.js?v={{ app_version }}"></script>
 <script src="/static/app.js?v={{ app_version }}"></script>
 </body>
 </html>
