+++ tests/contract_checks.ps1
@@ -112,8 +112,8 @@ Assert-Contains -Path "templates\index.html" -Pattern 'id="closeBackdrop"' -Mess
 Assert-Contains -Path "static\app.js" -Pattern 'socket\.on\(''show_close_dialog''' -Message "index.html must handle backend close confirmation requests"
 Assert-Contains -Path "static\app.js" -Pattern 'socket\.emit\(''close_choice''' -Message "index.html must send close choices to the backend"
 Assert-Contains -Path "templates\index.html" -Pattern 'id="alertBackdrop"' -Message "index.html must contain an app alert dialog"
-Assert-Contains -Path "static\app.js" -Pattern 'showAlertModal' -Message "index.html must show app alert dialogs from alert events"
-Assert-Contains -Path "static\app.js" -Pattern 'mergedAlertCount' -Message "alert dialogs must merge repeated alerts into one visible dialog"
+Assert-Contains -Path "static\alert-log-center.js" -Pattern 'showAlertModal' -Message "index.html must show app alert dialogs from alert events"
+Assert-Contains -Path "static\alert-log-center.js" -Pattern 'mergedAlertCount' -Message "alert dialogs must merge repeated alerts into one visible dialog"
 Assert-NotContains -Path "templates\index.html" -Pattern 'alertQueue' -Message "alert dialogs must not queue repeated modal popups"
 Assert-Contains -Path "static\app.css" -Pattern 'max-width:none' -Message "layout must use the available window width"
 Assert-Contains -Path "static\app.css" -Pattern 'grid-template-columns:minmax\(0, 1fr\) var\(--right-width\)' -Message "layout must use responsive grid columns"
@@ -378,16 +378,16 @@ Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("delete_watch_target"\)'
 Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("toggle_watch_target"\)' -Message "backend must toggle watch targets"
 Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("reset_watch_target"\)' -Message "backend must reset watch target trigger state"
 Assert-Contains -Path "static\alert-rule-center.js" -Pattern 'kind === ''watch_target''' -Message "unified alert center must render watch target rules"
-Assert-Contains -Path "static\app.js" -Pattern 'renderNotificationBadges' -Message "alert log must render notification delivery status"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("export_alert_log"\)' -Message "backend must export alert logs"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("clear_alert_log"\)' -Message "backend must clear alert logs"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("update_alert_log_status"\)' -Message "backend must update alert log read state"
-Assert-Contains -Path "app.py" -Pattern '@socketio\.on\("resend_alert_notification"\)' -Message "backend must resend alert notifications"
+Assert-Contains -Path "static\alert-log-center.js" -Pattern 'renderNotificationBadges' -Message "alert log must render notification delivery status"
+Assert-Contains -Path "goldmonitor\socket_alert_log.py" -Pattern '@socketio\.on\("export_alert_log"\)' -Message "backend must export alert logs"
+Assert-Contains -Path "goldmonitor\socket_alert_log.py" -Pattern '@socketio\.on\("clear_alert_log"\)' -Message "backend must clear alert logs"
+Assert-Contains -Path "goldmonitor\socket_alert_log.py" -Pattern '@socketio\.on\("update_alert_log_status"\)' -Message "backend must update alert log read state"
+Assert-Contains -Path "goldmonitor\socket_alert_log.py" -Pattern '@socketio\.on\("resend_alert_notification"\)' -Message "backend must resend alert notifications"
 Assert-NotContains -Path "templates\index.html" -Pattern 'alertLogModeTabs' -Message "frontend must not expose alert log range switching"
 Assert-Contains -Path "templates\index.html" -Pattern 'alertLogSearch' -Message "frontend must support alert log search"
 Assert-Contains -Path "templates\index.html" -Pattern 'alertUnreadCount' -Message "frontend must show unread alert counts"
 Assert-NotContains -Path "templates\index.html" -Pattern 'acknowledgeAlert' -Message "frontend must not expose alert acknowledgement"
-Assert-Contains -Path "static\app.js" -Pattern 'resendAlertNotification' -Message "frontend must support alert notification resend"
+Assert-Contains -Path "static\alert-log-center.js" -Pattern 'resendAlertNotification' -Message "frontend must support alert notification resend"
 Assert-Contains -Path "templates\index.html" -Pattern 'exportAlertLog' -Message "frontend must expose alert log export"
 Assert-Contains -Path "templates\index.html" -Pattern 'clearAlertLog' -Message "frontend must expose alert log clearing"
 Assert-NotContains -Path "templates\index.html" -Pattern 'class="threshold-section"' -Message "frontend must not keep the old duplicated threshold form"
