+++ tests/test_frontend_market_status_contract.py
@@ -13,7 +13,7 @@ def test_frontend_fetch_status_uses_degraded_status_contract():
 
 
 def test_frontend_alert_log_uses_notification_summary_contract():
-    js = Path("static/app.js").read_text(encoding="utf-8")
+    js = Path("static/alert-log-center.js").read_text(encoding="utf-8")
     start = js.index("function alertNotificationIssues")
     end = js.index("function buildLogEntry", start)
     body = js[start:end]
