+++ tests/test_run_checks_script.py
@@ -14,6 +14,7 @@ EXPECTED_COMPILE_TARGETS = (
     "goldmonitor/review_notes.py",
     "goldmonitor/scheduler.py",
     "goldmonitor/socket_alert_rules.py",
+    "goldmonitor/socket_alert_log.py",
     "goldmonitor/socket_history_review.py",
     "goldmonitor/socket_operations.py",
     "goldmonitor/socket_portfolio.py",
